+++ src/main/java/kitchenpos/ordertable/application/OrderOrderTableService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.ordertable.application;
 
 public interface OrderOrderTableService {
