+++ src/test/java/kitchenpos/ordertable/domain/OrderTablesTest.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.ordertable.domain;
 
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTables;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
