+++ src/test/java/kitchenpos/ui/TableGroupRestControllerTest.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import io.restassured.RestAssured;
 import kitchenpos.application.TableGroupService;
 import kitchenpos.common.controller.ControllerTest;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.repository.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.repository.OrderTableRepository;
 import kitchenpos.ui.dto.TableGroupCreateRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
